# swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Codecademy e-Commerce REST API",
      "description": "Defines the backend logic relating to the <strong>'registration'</strong>, <strong>'customers'</strong>, <strong>'addresses'</strong>, <strong>'orders'</strong>, <strong>'products'</strong>, <strong>'cart'</strong>, and <strong>'login'</strong> endpoints of an e-commerce REST API.  <code>GET</code> operations can be performed directly on the <strong>HEROKU</strong> database; <code>POST</code>/<code>PUT</code>/<code>DELETE</code> operations can be performed on the <strong>SWAGGER HUB</strong> <em>virtual mocking server</em> so that the <strong>HEROKU</strong> database is not compromised.",
      "version": "0.1"
    },
    "servers": [
      {
        "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
        "description": "Heroku App"
      },
      {
        "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
        "description": "SwaggerHub API Auto Mocking"
      }
    ],
    "paths": {
...
      "/api/customers": {
        "get": {
          "tags": [
            "Customers"
          ],
          "summary": "Return all customer information",
          "description": "Returns the 'id', 'first_name', 'last_name', 'username', 'password', and 'email' fields for all customers in the 'customers' table",
          "responses": {
            "200": {
              "description": "JSON array of customers",
              "content": {
                "application/json; charset=utf-8": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Customer"
                    }
                  },
                  "example": [
                    {
                      "id": 1,
                      "first_name": "Christopher",
                      "last_name": "Larham",
                      "username": "chris_larham_1983",
                      "password": "updated_chrissy_b0i*",
                      "email": "chrislarham@example.com"
                    },
                    {
                      "id": 2,
                      "first_name": "Charlotte",
                      "last_name": "Marie",
                      "username": "charlotte_marie_1986",
                      "password": "mar13p0os*",
                      "email": "charlottemarie@example.com"
                    },
                    {
                      "id": 3,
                      "first_name": "William",
                      "last_name": "Larham",
                      "username": "will_larham_2007",
                      "password": "w1lliam*",
                      "email": "williamlarham@example.com"
                    },
                    {
                      "id": 4,
                      "first_name": "Seth",
                      "last_name": "Larham",
                      "username": "seth_larham_2008",
                      "password": "s3th0*",
                      "email": "sethlarham@example.com"
                    },
                    {
                      "id": 5,
                      "first_name": "Alyssa",
                      "last_name": "Larham",
                      "username": "alyssa_larham_2010",
                      "password": "aly55a*",
                      "email": "alyssalarham@example.com"
                    },
                    {
                      "id": 6,
                      "first_name": "Roger",
                      "last_name": "Larham",
                      "username": "roger_larham_1955",
                      "password": "r0ger*",
                      "email": "rogerlarham@example.com"
                    },
                    {
                      "id": 7,
                      "first_name": "Judith",
                      "last_name": "Larham",
                      "username": "judi_larham_1956",
                      "password": "Jud1*",
                      "email": "judilarham@example.com"
                    },
                    {
                      "id": 8,
                      "first_name": "Matthew",
                      "last_name": "Rowlings",
                      "username": "matt_rowlings_1983",
                      "password": "Matti3*",
                      "email": "mattrowlings@example.com"
                    },
                    {
                      "id": 9,
                      "first_name": "Steen",
                      "last_name": "Jinks",
                      "username": "steen_jinks_1991",
                      "password": "Jink5*",
                      "email": "steenjinks@example.com"
                    },
                    {
                      "id": 10,
                      "first_name": "Joe",
                      "last_name": "Bloggs",
                      "username": "joe_bloggs_1992",
                      "password": "bl0ggs",
                      "email": "joebloggs@example.com"
                    },
                    {
                      "id": 11,
                      "first_name": "Lugali",
                      "last_name": "Bobtail",
                      "username": "lugali_bobtail_1758",
                      "password": "lugal1*",
                      "email": "lugalibobtail@example.com"
                    },
                    {
                      "id": 12,
                      "first_name": "Divine",
                      "last_name": "Intervention",
                      "username": "divine_intervention_0",
                      "password": "Genes1s*",
                      "email": "divineintervention@outlook.com"
                    },
                    {
                      "id": 13,
                      "first_name": "Customer First Name",
                      "last_name": "Customer Last Name",
                      "username": "Customer Username",
                      "password": "Customer Password",
                      "email": "Customer Email"
                    }
                  ]
                }
              }
            },
            "500": {
              "description": "Returns a description of the error {err} that was unexpectedly encountered",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "The unexpected value of {err.message}."
                  }
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
              "description": "Heroku App"
            },
            {
              "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
              "description": "SwaggerHub API Auto Mocking"
            }
          ]
        },
        "servers": [
          {
            "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
            "description": "Heroku App"
          },
          {
            "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
            "description": "SwaggerHub API Auto Mocking"
          }
        ]
      },
      "/api/customers/{customerId}": {
        "get": {
          "tags": [
            "Customers"
          ],
          "summary": "Return information relating to a specific customer",
          "description": "Returns the 'first_name', 'last_name', 'username', 'password', and 'email' fields relating to the customer whose id is 'customer_id'",
          "parameters": [
            {
              "name": "customerId",
              "in": "path",
              "description": "The customer ID",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "minimum": 1,
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "A customer object",
              "content": {
                "application/json; charset=utf-8": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  },
                  "example": "{\n  \"id\": 1,\n  \"first_name\": \"Christopher\",\n  \"last_name\": \"Larham\",\n  \"username\": \"chris_larham_1983\",\n  \"password\": \"updated_chrissy_b0i*\",\n  \"email\": \"chrislarham@example.com\"\n}  "
                }
              }
            },
            "404": {
              "description": "Customer with the specified ID not found in the database",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "Customer does not exist in the database."
                  }
                }
              }
            },
            "500": {
              "description": "Returns a description of the error {err} that was unexpectedly encountered",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "The unexpected value of {err.message}."
                  }
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
              "description": "Heroku App"
            },
            {
              "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
              "description": "SwaggerHub API Auto Mocking"
            }
          ]
        },
        "put": {
          "tags": [
            "Customers"
          ],
          "summary": "Update a specific customer's details",
          "description": "Updates the 'first_name', 'last_name', 'username', and 'password' details for the customer whose id is 'customer_id'",
          "parameters": [
            {
              "name": "customerId",
              "in": "path",
              "description": "The customer ID",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "minimum": 1,
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerUpdate"
                },
                "examples": {
                  "0": {
                    "value": "{\n    \"first_name\": \"Christopher\",\n    \"last_name\": \"Larham\",\n    \"username\": \"chris_larham_1983\",\n    \"password\": \"updated_chrissy_b0i*\",\n    \"email\": \"chrislarham@example.com\"\n}"
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "String confirmation that the customer's details were updated successfully",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "Customer with id {customerId} updated successfully"
                  }
                }
              }
            },
            "404": {
              "description": "Customer with the specified ID not found in the database",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "Customer does not exist in the database."
                  }
                }
              }
            },
            "500": {
              "description": "Returns a description of the error {err} that was unexpectedly encountered",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "The unexpected value of {err.message}."
                  }
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
              "description": "SwaggerHub API Auto Mocking"
            }
          ]
        },
        "delete": {
          "tags": [
            "Customers"
          ],
          "summary": "Delete a specific customer from the database",
          "description": "Deletes the customer details relating to the customer whose id is 'customerId' from the 'customers' table",
          "parameters": [
            {
              "name": "customerId",
              "in": "path",
              "description": "The customer ID",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "minimum": 1,
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "String confirmation that the customer details belonging to the customer with id 'customerId' have been deleted from the 'customers' table",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string"
                  },
                  "example": "Customer with id {customerId} deleted successfully"
                }
              }
            },
            "500": {
              "description": "Returns a description of the error {err} that was unexpectedly encountered",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "The unexpected value of {err.message}."
                  }
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
              "description": "SwaggerHub API Auto Mocking"
            }
          ]
        },
        "servers": [
          {
            "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
            "description": "SwaggerHub API Auto Mocking"
          }
        ]
      },
...
    },
    "components": {
      "schemas": {
...
        "Customer": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "description": "The customer's ID"
            },
            "first_name": {
              "type": "string",
              "description": "The customer's first name"
            },
            "last_name": {
              "type": "string",
              "description": "The customer's last name"
            },
            "username": {
              "type": "string",
              "description": "The customer's username"
            },
            "password": {
              "type": "string",
              "description": "The customer's password"
            },
            "email": {
              "type": "string",
              "description": "The customer's email address"
            }
          }
        },
        "CustomerUpdate": {
          "type": "object",
          "properties": {
            "first_name": {
              "type": "string",
              "description": "The customer's first name"
            },
            "last_name": {
              "type": "string",
              "description": "The customer's last name"
            },
            "username": {
              "type": "string",
              "description": "The customer's username"
            },
            "password": {
              "type": "string",
              "description": "The customer's password"
            },
            "email": {
              "type": "string",
              "description": "The customer's email address"
            }
          }
        },
...
      }
    }
  }